Reimplement the savegame encoder as package rather than a single module
The new implementation is both cleaner and significantly faster than the previous
implementation
